requirements.txt
PySide6
imagehash
numpy
pillow
loguru
pyinstaller
chromadb
sentence-transformers
transformers
torch
torchvision
humanize
pydantic
